API redesign, closes #38
Since SleeperNotifier does not need to be thread-safe once it reaches its destination thread, we were paying a performance penalty on every usage of it.
Instead, a thread-safe SleeperHandlerEntry can be sent to the destination thread, which can then be used to create the notifier instance on arrival.

In addition, addNotifier() no longer accepts a SleeperNotifier instance, and now returns a SleeperHandlerEntry which must be used.

I'm not completely sold on this API yet, since SleeperHandlerEntry is not actually referenced anywhere within SleeperHandler, but it's definitely an improvement.